fix: duplicate txs in transaction history
The reason: sometimes `fetchMoreTransactions` is being called multiple times for the same paging_token.
This is likely due to the fact that `fetchMoreTransactions` is being called in `useEffect` and thus may be called multiple times (e.g. on rerender).
This is the problem on it is own (excessive requests). However, it was possible that when called twice for the same `paging_token`, the first request
updates the cache with new txs and then second requests tries to add the same set of txs to the cache — thus duplicates.

Easy fix is to use local-scoped `prevTransactions` when updating cache — this way the cache will be rewritten, no duplicates. The `inefficiency of multiple requests
still stays in the code though, just not visible.`

Proper fix would be to make sure `fetchMoreTransactions` is being called just once per `paging_token`